trading/order_book_snapshot/pyproject.toml
[project]
name = "order-book-snapshot"
version = "0.1.0"
description = "Live asset order book snapshot service"
requires-python = ">=3.11"

dependencies = [
  "tornado>=6.4",
  "asyncpg>=0.29",
  "httpx>=0.27",
  "python_dotenv>=1.2.0",
  "pydantic>=2.12.2"
]
